+++ lib/widgets/buttons/icon_button.dart
@@ -3,11 +3,11 @@ import '../../functions/design.dart';
 
 /// Use UdIconButton when you want only an Icon widget to work like a button. You can pass Icon widget, padding arround it and a function to execute on tap.
 class UdIconButton extends StatelessWidget {
-  final Icon icon;
+  final Icon child;
   final EdgeInsets padding;
   final Function function;
 
-  UdIconButton({this.icon, this.padding, this.function});
+  UdIconButton({this.child, this.padding, this.function});
   @override
   Widget build(BuildContext context) {
     double _design = doNotUseThisDesignValue(context: context);
@@ -17,7 +17,7 @@ class UdIconButton extends StatelessWidget {
       focusColor: Colors.transparent,
       child: Padding(
         padding: padding ?? EdgeInsets.all(_design * 0.5),
-        child: icon,
+        child: child,
       ),
       onTap: function,
     );